Refactoring changes: split the files into separate ones based on the supported datasets
